Moving pollStop to reusable lib function
