+++ src/routes/metro/[stop]/+layout.server.ts
@@ -1,10 +1,12 @@
 import type { stopDB } from '$lib/stopDB';
+import { pollStop } from '$lib/stop';
 import type { LayoutServerLoad } from './$types';
 import type { Collection } from "mongodb";
+import { error } from '@sveltejs/kit';
 
 //DB data is returned first, passed to the client load page. This never changes, so it shouldn't expire
 export const load: LayoutServerLoad = async ({ params, locals, depends }) => {
-    depends('stopDB')
+    depends('metro')
 
     const { stops }: { stops: Collection<stopDB> } = locals;
     const code = parseInt(params.stop);
@@ -13,6 +15,20 @@ export const load: LayoutServerLoad = async ({ params, locals, depends }) => {
 
     return {
         code,
-        db: res
+        db: res,
+        api: { promise: getStop(code) }
     };
+}
+
+//Wraps pollStop and adds error checks
+async function getStop(code: number) {
+    try {
+        return await pollStop(code);
+    } catch (e) {
+        if (e === 'noInfo') {
+            throw error(404, { message: 'No information available' });
+        } else {
+            throw error(500, { message: 'GTT API is offline' });
+        }
+    }
 }
